Split mazebuilder in different files
